Fixed K3bDevice::Device::getNextWritableAddress. made the track argument of readTrackInformation an int instead of a unsigned long.
Moved lib version to 4 hoping this were the last bin-comp braking changes.

svn path=/trunk/extragear/multimedia/k3b/; revision=545858

+++ libk3bdevice/k3bdevice.h
@@ -547,7 +547,7 @@ namespace K3bDevice
        *        \li 10b - value refers to a session number
        *
        */
-      bool readTrackInformation( unsigned char** data, unsigned int& dataLen, int type, unsigned long value ) const;
+      bool readTrackInformation( unsigned char** data, unsigned int& dataLen, int type, int value ) const;
 
       /**
        * if true is returned dataLen specifies the actual length of *data which needs to be
